flt22

+++ Entrega2/promedio.sql
@@ -1,6 +1,6 @@
 CREATE OR REPLACE FUNCTION
 promedio(t1 date, t2 date, t3 varchar(50))
-RETURNS TABLE (iid int, promedio NUMERIC) AS $$
+RETURNS TABLE (iid int, promedio FLOAT) AS $$
 DECLARE
 	tupla RECORD;
 	tupla2 RECORD;
